Adjusting testing setup to support server-side tests and coverage
