+++ app/package.json
@@ -30,10 +30,13 @@
   "scripts": {
     "client": "cd client && npm start",
     "server": "cd server && npm start",
+    "server_coverage": "cd server && npm run start_coverage",
     "start": "concurrently -kc \"blue.dim,green.dim\" \"npm:server\" \"npm:client\"",
     "setup": "echo \"Installing project dependencies...\" && npm ci --legacy-peer-deps && echo \"Installing client dependencies...\" && cd client && npm ci --legacy-peer-deps && echo \"Installing server dependencies...\" && cd .. && cd server && npm ci",
     "cypress": "concurrently -kc \"blue.dim,green.dim,yellow.dim\" -n server,client,cypress \"npm:server\" \"npm:client\" \"cypress open --env coverage=false\"",
-    "coverage": "concurrently -kc \"blue.dim,green.dim,yellow.dim\" -n server,client,cypress \"npm:server\" \"npm:client\" \"npx rimraf .nyc_output/out.json && cypress run --browser electron\"",
+    "coverage": "npx rimraf .nyc_output/out.json && npx rimraf coverage && concurrently -kc \"blue.dim,green.dim,yellow.dim\" -n server,client,cypress \"npm:server_coverage\" \"npm:client\" \"cypress run --browser electron\"",
+    "coverage_spec_ex": "npx rimraf .nyc_output/out.json && concurrently -kc \"blue.dim,green.dim,yellow.dim\" -n server,client,cypress \"npm:server_coverage\" \"npm:client\" \"cypress run --browser electron --spec 'cypress/e2e/state.cy.ts'\"",
+    "full_coverage": "cd server && npm run test && cd .. && npm run coverage",
     "glossary": "cd server && npm run glossary"
   },
   "devDependencies": {
