+++ app/server/app/server/routes/proxy.js
@@ -10,7 +10,7 @@ module.exports = function (app) {
   const router = express.Router();
 
   // only expose proxy in local environment
-  if (!app.enabled('isLocal')) return;
+  if (!app.enabled('isLocal') || !app.enabled('isTest')) return;
 
   router.get('/', function (req, res, next) {
     let authoriztedURL = false;
