+++ app/server/app/server/utilities/environment.js
@@ -1,12 +1,14 @@
 // determine which environment we are in
 exports.getEnvironment = function () {
   let isLocal = false;
+  let isTest = false;
   let isDevelopment = false;
   let isStaging = false;
   let isProduction = false;
 
   if (process.env.NODE_ENV) {
     isLocal = 'local' === process.env.NODE_ENV.toLowerCase();
+    isTest = 'test' === process.env.NODE_ENV.toLowerCase();
     isDevelopment = 'development' === process.env.NODE_ENV.toLowerCase();
     isStaging = 'staging' === process.env.NODE_ENV.toLowerCase();
     isProduction = 'production' === process.env.NODE_ENV.toLowerCase();
@@ -14,6 +16,7 @@ exports.getEnvironment = function () {
 
   return {
     isLocal,
+    isTest,
     isDevelopment,
     isStaging,
     isProduction,
