+++ app/server/jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+  setupFiles: ['./jest_setup.js'],
+  roots: [
+    "tests"
+  ],
+  testPathIgnorePatterns: [
+    "node_modules/",
+    "dist/",
+    "build/",
+    ".nyc_output/",
+    "coverage/"
+  ],
+};
