+++ app/server/package.json
@@ -30,6 +30,8 @@
   "main": "app/server.js",
   "scripts": {
     "start": "nodemon -r dotenv/config --inspect app/server.js dotenv_config_path=./.env.local",
+    "start_coverage": "nyc --silent node --env-file=./.env.local app/server.js",
+    "test": "npx rimraf .nyc_output/out.json && npx rimraf coverage && jest --coverage --coverageDirectory='coverage'",
     "glossary": "node -r dotenv/config --inspect app/tasks/updateGlossary.js dotenv_config_path=./.env.local"
   },
   "dependencies": {
@@ -44,11 +46,15 @@
     "serve-favicon": "2.5.0"
   },
   "devDependencies": {
+    "@cypress/code-coverage": "^3.12.48",
     "browser-sync": "3.0.2",
     "husky": "9.1.5",
+    "jest": "^29.7.0",
     "lint-staged": "15.2.9",
     "nodemon": "3.1.4",
-    "prettier": "3.3.3"
+    "nyc": "^17.0.0",
+    "prettier": "3.3.3",
+    "supertest": "^7.0.0"
   },
   "husky": {
     "hooks": {
