+++ app/server/tests/api.health.test.js
@@ -1,0 +1,13 @@
+const supertest = require("supertest");
+const app = require("../app/app");
+
+describe("Test example", () => {
+  test("GET /api/health should return UP", async () => {
+    const response = await supertest(app)
+      .get("/api/health")
+      .expect(200)
+      .expect("Content-Type", /json/);
+
+    expect(response.body).toEqual({ status: "UP" });
+  });
+});
